--- a/go.mod
+++ b/go.mod
@@ -1,5 +1,3 @@
 module github.com/ianlopshire/go-fixedwidth
 
 go 1.13
-
-require github.com/pkg/errors v0.8.1